Refactor content ID types to long, update date format in requests/responses, and enhance upload file handling

+++ src/main/java/vn/fu_ohayo/dto/response/ContentSpeakingResponse.java
@@ -16,7 +16,7 @@ import java.util.Date;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ContentSpeakingResponse {
 
-    private int contentSpeakingId;
+    private long contentSpeakingId;
     private String title;
     private String image;
     private CategorySpeakingEnum category;
